requirements.txt
flask
flask-cors
langchain
langchain-openai
langchain-groq
langchain-community
langgraph
python-dotenv
streamlit
requests
